added addresing modes + refactor

+++ app.go
@@ -5,8 +5,7 @@ import "cpu/nes"
 func main() {
 	println("hello world")
 
-	var program = []int{0x0000}
+	var program = []uint8{0x0000}
 
 	nes.ProgramLoop(program)
-
 }
